requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
google-genai==1.30.0
openai==1.99.1
python-docx==1.2.0
pypdf==5.9.0
pdfplumber==0.11.7
Pillow==11.3.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytest==8.4.1
httpx==0.28.1
